module 02 done

+++ cpp_module_02/ex01/Fixed.cpp
@@ -6,74 +6,87 @@
 /*   By: oel-yous <oel-yous@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/27 14:30:38 by oel-yous          #+#    #+#             */
-/*   Updated: 2021/09/27 16:24:38 by oel-yous         ###   ########.fr       */
+/*   Updated: 2021/09/29 14:23:33 by oel-yous         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "Fixed.hpp"
 
-Fixed::Fixed( void ) : _fixedPointvalue(0)
-{
+// ------------------ Constructors ----------------------//
+
+Fixed::Fixed( void ) : _fixedPointvalue(0){
     std::cout << "Default Constructor called" << std::endl;
     return ;
 }
 
-Fixed::Fixed( Fixed const & src)
-{
+Fixed::Fixed( Fixed const & src){
     std::cout << "Copy constructor called" << std::endl;
     *this = src;
     return ;
 }
 
-Fixed::Fixed(const int i) 
-{
+Fixed::Fixed(const int i) {
     this->_fixedPointvalue = i << this->_fractionalBits;
     std::cout << "Int constructor called" << std::endl;
     return ;
 }
 
-Fixed::Fixed(const float f)
-{
+Fixed::Fixed(const float f){
     this->_fixedPointvalue = roundf(f * (1 << this->_fractionalBits));
     std::cout << "Float constructor called" << std::endl;
     return ;
 }
 
-Fixed::~Fixed()
-{
+// -------------------------------------------------------//
+
+// ------------------ Destructor -------------------------//
+
+Fixed::~Fixed(){
     std::cout << "Destructor called" << std::endl;
     return ;
 }
 
-Fixed & Fixed::operator=( Fixed const & rhs )
-{
+// -------------------------------------------------------//
+
+// ------------ assignation operator overload -------------//
+
+Fixed & Fixed::operator=( Fixed const & rhs ){
     std::cout << "Assignation operator called " << std::endl;
     this->_fixedPointvalue = rhs.getRawBits();
     return *this;
 }
 
-int Fixed::getRawBits( void ) const
-{
+// -------------------------------------------------------//
+
+// ----------------- setter and getter  ------------------//
+
+int Fixed::getRawBits( void ) const{
     return (this->_fixedPointvalue);
 }
 
-void Fixed::setRawBits(int const raw)
-{
+void Fixed::setRawBits(int const raw){
     this->_fixedPointvalue = raw;
 }
 
-float Fixed::toFloat( void ) const
-{
-    return (((float)(this->_fixedPointvalue) / (float)(1 << this->_fractionalBits)));
+// -------------------------------------------------------//
+
+// --------------- convert to float/to int  ---------------//
+
+float Fixed::toFloat( void ) const{
+    return (((float)(this->_fixedPointvalue) / (1 << this->_fractionalBits)));
 }
 
-int Fixed::toInt( void ) const
-{
+int Fixed::toInt( void ) const{
     return (this->_fixedPointvalue / (1 << this->_fractionalBits));
 }
 
-std::ostream & operator<<(std::ostream & o, Fixed const & rhs)
-{
+// -------------------------------------------------------//
+
+// ------------- overload to the « operator  --------------//
+
+std::ostream & operator<<(std::ostream & o, Fixed const & rhs){
     o << rhs.toFloat();
     return o;
 }
+
+// -------------------------------------------------------//
